third_party/blink/web_tests/wpt_internal/overscroll-gestures/scroll-into-view.html

<!doctype html>
<!DOCTYPE html>
<html >
<title>Overscroll position: scrollIntoView on content</title>
<meta name="author" content="github.com/vmpstr">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
html,body {
  margin: 0;
  padding: 0;
}
.spacer {
  height: 1000px;
}
#target {
  width: 50px;
  height: 50px;
  background: green;
}
#overscroll {
  top: -1000px;
  height: 1px;
  width: 100%;
}
.scroller {
  overflow: scroll;
  width: 300px;
  height: 300px;
}
</style>

<div class=scroller overscrollcontainer>
  <div class=spacer></div>
  <div id=target></div>
  <div class=spacer></div>
  <div id=overscroll></div>
</div>

<button command="toggle-overscroll" commandfor="overscroll">toggle</button>

<script>
test(() => {
  assert_equals(target.getBoundingClientRect().y, 1000, "before");
  target.scrollIntoView();
  assert_equals(target.getBoundingClientRect().y, 0, "after");
}, "scrollIntoView on content");
</script>
